Several changes have been made - Moved the interfaces to a folder with the name interfaces. - An interface for typing the S3 workspace has been created. - Modified the filemanager service requests to build the headers correctly.
