+++ projects/ontimize-web-ngx-filemanager/src/lib/interfaces/workspace.provider.interface.ts
