+++ projects/ontimize-web-ngx-filemanager/src/lib/interfaces/workspaceS3.interface.ts
@@ -1,0 +1,4 @@
+export interface WorkspaceS3{
+    name: string;
+    data: any;
+}
